o2c-graph-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
networkx==3.2.1
pandas==2.1.3
google-generativeai==0.3.0
python-dotenv==1.0.0
